+++ CHANGELOG.md
@@ -1,5 +1,3 @@
-# v2.12.1
-
-* bugfix release: fixed a bug with `stringify`
+# v2.12.2